RM_Partner/Rm_bot_maid/tst.py
import asyncio
import logging
from telegram import Update
from telegram.ext import ApplicationBuilder, CommandHandler, ContextTypes

logging.basicConfig(
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    level=logging.INFO
)
logger = logging.getLogger(__name__)

TOKEN = "8008686085:AAE3Iq1YJhWpksbo0yeuxfFgNw_4DDWH6w0"

async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
    await update.message.reply_text("Bot started!")

async def main():
    app = ApplicationBuilder().token(TOKEN).build()
    app.add_handler(CommandHandler("start", start))
    await app.run_polling(close_loop=False)

if __name__ == '__main__':
    # Simply use asyncio.run() — do NOT mix get_running_loop() here
    asyncio.run(main())
