requirements.txt
anyio==4.9.0
certifi==2025.4.26
colorama==0.4.6
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-dotenv==1.1.1
python-telegram-bot==22.1
sniffio==1.3.1
pyrogram==2.0.99
tgcrypto
telethon